Wrap point-in-time calculation state in a Calculation class.
There is no ambiguity of which point in time is represented by the
data in the gmx::restraint::Calculation class. Unfortunately, it will
have side effects if calculation updates some of the output parameters
before an exception is thrown, so we should revisit.

Attempts are being made to clarify data that is input, output, or
modified in place, but require factoring of other aspects of simulation
state.

The intent of the class is to replace pull_potential().