gmxapi-62 Stop condition hook tasks 2 and 3
Allow GROMACS client code to set the stop condition during the
integrator execution.

Provide a gmxapi hook to let client code issue a data event that causes
the framework to set the stop condition.

Adds gmxapi::md::signals enum, gmxapi::Signal proxy class (functor),
and gmxapi::getMdrunnerSignal helper function, all of which are more
proof of concept than on track for gmxapi specification. This is to
test out an evolving facility for plugins (currently via the
IRestraintPotential::bindSession hook and accompanying experimentation
in the sample_restraint) for workflow resources to provide plugin code
with a to set the SimulationSignals used by the integrator.