Beginnings of SessionResources.
We're running into problems with our inability to map interactions
between workflow members to specific unique elements. This is
particularly problematic when trying to interpret logical operations on
the outputs of multiple bound restraints.

These changes introduce a SessionResources class to represent the
distinct interface of each distinct workflow element with the single
session. The binding interface is updated and the signalling mechanism
is reimplemented. An additional header gmxapi/session/resources.h
provides only a forward declaration of SessionResources, which is right
now used as a handle in free functions that serve as its public
interface.

A SignalManager class provides an abstraction to the signal inputs to
a gmx::Mdrunner instance.