Initial commit for a sandbox branch.
Acts as a patch for GROMACS master to serve as a base for Kasson Lab
fork. Feature branches based on the Kasson Lab fork are easily converted
to Gerrit changes based by rebasing from their fork point to master.

Not targeted for merging with master. Could serve as a sandbox branch
in Gerrit.

Change-Id: Ic74ee9abc70f9d1be3e1d94619477c65ee093cb0