Reduce logger line length.
Extra elements had crept into the logger hierarchy. We probably want to
keep the gmxapi import of the Python logging module hidden. If we move
to a different logging facility in the future, we don't want to have to
chase down modules that are using the the Python built-in module. But
we may still prefer to normalize to getLogger() instead of getChild() in
the future and just have each gmxapi module do getLogger(__name__).