Don't install sample restraint package by default.
When building the sample_restraint project as part of a GROMACS build,
do not installed the myplugin Python package as part of the `install`
target. This is almost undoubtedly not what the user intended.

However, the package is still built when GMX_PYTHON_PACKAGE=ON to help
catch errors and facilitate testing by developers or automated systems
(i.e. Jenkins).