Remove OperationDetails instance from ResourceManager.
Even though OperationDetails and ResourceManager are both implementation
details of the gmxapi.operation module Context, the coupling between
the classes is confusing and burdensome. Removing the OperationDetails
instance from ResourceManager is a step along the way of simplifying
the definitions of Operations, clarifying the Context NodeBuilder
interface, and clarifying / improving the ownership and lifetime
management semantics of operation instances and their resources.

Additionally, move the remaining use of OperationDetails from the
NodeBuilder ABC to the constructor of the builders that use it.